requirements.txt
pyperclip==1.7.0
PyAutoGUI==0.9.52
PyMySQL==0.9.3
openpyxl==3.0.0
Pillow==8.3.1
PyQt5==5.15.4
SQLAlchemy==1.4.22
